% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\name{get_crypto_meta}
\alias{get_crypto_meta}
\title{Get static metdata}
\usage{
get_crypto_meta(apikey, symbol = NULL, id = NULL, slug = NULL,
  test = FALSE)
}
\arguments{
\item{apikey}{A valid API-key from Coinmarketcap}

\item{symbol}{One or more cryptocurrency symbols.
Example: c("BTC","ETH").}

\item{id}{Alternatively pass one or more CoinMarketCap cryptocurrency IDs.
Example: c(1,2)}

\item{slug}{Alternatively pass a vector of cryptocurrency slugs.
Example: c("bitcoin","ethereum")}

\item{test}{If `TRUE`, the requests are done in a testing sandbox environment.
This requires an extra testing API key, but allows possible all requests.}
}
\value{
A dataframe with metadata of Cryptocurrencies
}
\description{
Get all static metdata available for one or more cryptocurrencies
}
\note{
At least one "id" or "slug" or "symbol" is required for this request.
}
\examples{
\dontrun{
get_crypto_meta(apikey)
get_crypto_meta(apikey, symbol = c("BTC","ETH"))
get_crypto_meta(apikey, id = c(1,2,3,4))
get_crypto_meta(apikey, slug=c("bitcoin", "ethereum"))
}
}
\references{
\href{https://coinmarketcap.com/api/documentation/v1/#operation/getV1CryptocurrencyInfo}{API documentation}
}
