% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchanges.R
\name{get_exchange_meta}
\alias{get_exchange_meta}
\title{Get all cryptocurrency exchanges metadata}
\usage{
get_exchange_meta(apikey, id = NULL, slug = NULL, test = FALSE)
}
\arguments{
\item{apikey}{A valid API-key from Coinmarketcap}

\item{test}{If `TRUE`, the requests are done in a testing sandbox environment.
This requires an extra testing API key, but allows possible all requests.}
}
\value{
A dataframe with exchange metadata values
}
\description{
Returns all static metadata for one or more exchanges.
This information includes details like launch date, logo,
official website URL, social links, and market fee documentation URL.
}
\examples{
\dontrun{
get_exchange_meta(apikey)
get_exchange_meta(apikey, id = 1)
get_exchange_meta(apikey, slug = c("binance", "gdax"))
}
}
\references{
\href{https://coinmarketcap.com/api/documentation/v1/#operation/getV1ExchangeInfo}{API documentation}
}
