% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currencies.R
\name{get_crypto_map}
\alias{get_crypto_map}
\title{Get active cryptocurrencies}
\usage{
get_crypto_map(apikey, test = FALSE, ...)
}
\arguments{
\item{apikey}{A valid API-key from Coinmarketcap}

\item{test}{If `TRUE`, the requests are done in a testing sandbox environment.
This requires an extra testing API key, but allows possible all requests.}

\item{...}{Further arguments passed to the request. Further information
can be found in the \href{https://coinmarketcap.com/api/documentation/v1/#operation}{API documentation}}
}
\value{
A dataframe with all active cryptocurrencies supported by the
platform including a unique id for each cryptocurrency.
}
\description{
Get all active cryptocurrencies supported by the
platform including a unique id
}
\examples{
\dontrun{
get_crypto_map(apikey)
get_crypto_map(apikey, symbol="BTC")
get_crypto_map(apikey, symbol=c("BTC","ETH"))
get_crypto_map(apikey, listing_status = "active", start = 1, limit = 10)
get_crypto_map(apikey, listing_status = "inactive", start = 1, limit = 10)
}
}
\references{
\href{https://coinmarketcap.com/api/documentation/v1/#operation/getV1CryptocurrencyMap}{API documentation}
}
